+++ example/razzle-ssr/src/client.js
@@ -3,10 +3,19 @@ import BrowserRouter from 'react-router-dom/BrowserRouter';
 import React from 'react';
 import { render } from 'react-dom';
 
+import { I18nextProvider } from 'react-i18next';
+import i18n from './i18n';
+
 render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
+  <I18nextProvider
+    i18n={ i18n }
+    initialI18nStore={window.initialI18nStore}
+    initialLanguage={window.initialLanguage}
+  >
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </I18nextProvider>,
   document.getElementById('root')
 );